+++ src/components/Controls/styles.js
@@ -1,0 +1,38 @@
+import * as v from '../../styles/variables'
+
+const styles = {
+  control: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'baseline',
+    alignContent: 'space-between',
+    flexWrap: 'wrap'
+  },
+  controlLeft: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'baseline',
+    alignContent: 'space-between',
+    flexWrap: 'wrap'
+  },
+  controlRight: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'baseline',
+    alignContent: 'space-between'
+  },
+  controlPiece: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'baseline',
+    alignContent: 'space-between',
+    margin: `${v.lnht}rem`,
+    verticalAlign: 'middle'
+  }
+}
+
+export default styles