+++ src/components/Results/resultsStyles.js
@@ -1,0 +1,15 @@
+import * as v from '../../styles/variables'
+
+const styles = {
+  results: {
+    marginTop: `${v.lnht}rem`,
+    paddingTop: `${v.lnht}rem`,
+    borderTop: {
+      width: 1,
+      style: 'dotted',
+      color: v.bluNtl.string()
+    }
+  }
+}
+
+export default styles