+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 
 - Fix PropTypes for children in Button and Help files.
 - Create a Form component with shared styles and refactor all tool forms to use the new component.
+- Create a Results component with shared styles and refactor all tool results to use the new component.
+- Create various Controls components and refactor all tool forms and results to use the new components.
 - Update dependencies.
 
 ### Removed 0.2.3